requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.0.2
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.3.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
billiard==3.6.3.0
bleach==6.0.0
celery==5.0.5
certifi==2020.12.5
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.6
comm==0.1.3
confusable-homoglyphs==3.2.0
dataclasses-json==0.5.8
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==3.1.5
django-celery-results==2.5.1
django-extensions==3.2.3
django-registration==3.3
django-rest-framework==0.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
eventlet==0.33.3
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.17.1
filelock==3.12.2
fqdn==1.5.1
frozenlist==1.3.3
greenlet==2.0.2
idna==2.10
importlib-metadata==6.6.0
importlib-resources==5.12.0
install==1.3.5
ipykernel==6.23.2
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.1
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kombu==5.0.2
langchain==0.0.200
langchainplus-sdk==0.0.10
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mistune==2.0.5
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.2.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.5.0
nbformat==5.9.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.3
numexpr==2.8.4
numpy==1.24.3
openai==0.27.7
openapi-schema-pydantic==1.2.4
overrides==7.3.1
packaging==23.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==8.3.1
pkgutil_resolve_name==1.3.10
platformdirs==3.5.3
prometheus-client==0.17.0
prompt-toolkit==3.0.10
psutil==5.9.5
pure-eval==0.2.2
pycparser==2.21
pycryptodome==3.18.0
pycryptodomex==3.18.0
pydantic==1.10.9
Pygments==2.15.1
PyJWT==2.7.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2020.5
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
redis==3.5.3
requests==2.25.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
Send2Trash==1.8.2
six==1.15.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.16
sqlparse==0.4.1
stack-data==0.6.2
tenacity==8.2.2
terminado==0.17.1
tinycss2==1.2.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.2
vine==5.0.0
virtualenv==20.23.1
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.3
widgetsnbextension==4.0.7
xmltodict==0.13.0
yarl==1.9.2
zipp==3.15.0
